Document feature flag overrides and fix community post params
