+++ prompt_to_resume.txt
@@ -5,25 +5,27 @@
 - Fixed the MDX syntax issue in `content/curriculum/T3_Advanced/A-UVM-1_Advanced_Sequencing/index.mdx` so markdown parsing no longer fails ([D2]).
 - Documented local setup in the README, added Firebase placeholders to `.env.example`, and wired `npx playwright install` into `postinstall` to keep E2E tests runnable ([D3]).
 - Re-organized the `docs` directory to create a clear separation between `planning` and `feedback` documents.
+- Completed Milestone 1 learning-first foundations: filled out the Learning-First audit for every route ([L1]), introduced centralized feature flags ([L2]), and gated community/dashboard/history/account surfaces behind default-off flags ([L3]).
 
 **Current Status**
 - Milestone 3 (Stabilize) is complete: curriculum pages render again and the local dev checklist is up to date.
-- The `docs` directory has been restructured for better organization. Planning and feedback documents are now in `docs/planning` and `docs/feedback` respectively.
+- Milestone 1 learning-first gating work is also complete: feature flags ship in `src/tools/featureFlags.ts`, mock routes now 404/placeholder when disabled, and the audit doc calls out ownership per flag.
+- Feature flags can be toggled locally by setting `NEXT_PUBLIC_FEATURE_FLAG_<FLAG_NAME>` environment variables (accepts `true/false` or `1/0`).
 - `npm run type-check` passes; Playwright binaries install automatically during `npm install`.
 - E2E tests still fail due to brittle locators (`strict mode violations`)—addressed later in Milestone 4.
 
 **Next Objectives**
 1. Begin Milestone 1 (Learning-First & Scaffolding):
-   - [L1] Build `docs/feedback/Learning-First-Audit.md` by auditing every route listed in the Page & Route Inventory.
-   - [L2] Introduce `src/tools/featureFlags.ts` with the specified default-off flags and an `isFeatureEnabled` helper.
-   - [L3] Gate mock UI routes/components behind the new flags (e.g., dashboard, community, history) returning placeholders or `notFound()` when disabled.
-2. After L1–L3, proceed to Milestone 1 maintenance tasks [M1-1] and [M1-2] (enforce TS/ESLint in builds) before touching Milestone 2 or 4 work.
+   - ✅ [L1] Build `docs/feedback/Learning-First-Audit.md` by auditing every route listed in the Page & Route Inventory.
+   - ✅ [L2] Introduce `src/tools/featureFlags.ts` with the specified default-off flags and an `isFeatureEnabled` helper.
+   - ✅ [L3] Gate mock UI routes/components behind the new flags (e.g., dashboard, community, history) returning placeholders or `notFound()` when disabled.
+2. Next up: tackle Milestone 1 maintenance tasks [M1-1] and [M1-2] (remove `ignoreBuildErrors`/`ignoreDuringBuilds` and address resulting TS/ESLint violations) before moving to Milestone 2 or 4.
 
 **How to Get Started**
 - Read `docs/planning/Actionable-Task-Plan.md` for the step-by-step backlog.
 - Refer to the `docs/feedback` directory for analysis and audit reports.
 - Verify the dev environment: `npm install` (installs Playwright) → `npm run dev` to spot regressions → `npm run type-check`.
-- When tackling [L3], be methodical—introduce feature flags first, then update pages/components, and finally confirm routing behavior in the browser.
+- When tackling [M1-1]/[M1-2], remove the `ignoreBuildErrors`/`ignoreDuringBuilds` guards first, run `npm run type-check` and `npm run lint`, then iterate on the reported issues before re-running `npm run build`.
 
 **Known Risks / Follow-ups**
 - Do not resume feature roadmap items (interactive placement quiz, notifications, settings persistence) until after Milestones 1–4 are satisfied.
