+++ src/app/community/test-page/page.tsx
@@ -1,42 +1,11 @@
-import React from 'react';
-import dynamic from 'next/dynamic';
-import { type ExplanationStep } from '@/components/ui/InteractiveCode';
+import { notFound } from 'next/navigation';
+import { isFeatureEnabled } from '@/tools/featureFlags';
 
-// Disable server-side rendering for the interactive editor which relies on `window`.
-const InteractiveCode = dynamic(
-  () => import('@/components/ui/InteractiveCode'),
-  { ssr: false }
-);
+export default async function TestPage() {
+  if (!isFeatureEnabled('community')) {
+    notFound();
+  }
 
-const testCode = 
-'// test_sequence.sv\n' +
-'class test_sequence extends uvm_sequence;\n' +
-'  `uvm_object_utils(test_sequence)\n' +
-'\n' +
-'  function new(string name="test_sequence", uvm_object parent=null);\n' +
-'    super.new(name, parent);\n' +
-'  endfunction\n' +
-'\n' +
-'  virtual task body();\n' +
-'    `uvm_info("SEQ", "Starting test sequence", UVM_MEDIUM)\n' +
-'  endtask\n' +
-'endclass'.trim();
-
-const testExplanationSteps: ExplanationStep[] = [
-  { target: "2-11", title: "Test Sequence", explanation: "A simple test sequence." },
-];
-
-export default function TestPage() {
-  return (
-    <div className="p-8">
-      <h1 className="text-2xl font-bold mb-4">Component Test Page</h1>
-      <InteractiveCode
-        language="systemverilog"
-        explanationSteps={testExplanationSteps}
-        fileName="test.sv"
-      >
-        {testCode}
-      </InteractiveCode>
-    </div>
-  );
+  const TestPageClient = (await import('./TestPageClient')).default;
+  return <TestPageClient />;
 }
